ticket:26 Moved tests to separate folder
